+++ docs/versionhistory.rst
@@ -15,6 +15,7 @@ This library adheres to `Semantic Versioning <http://semver.org/>`_.
 - Fixed error during ``WAMPClient.close()`` if a connection attempt was in progress
 - Added compatibility with Asphalt 4.0
 - Added the ``concurrency`` option for procedure registration
+- Added the ``get_retained`` option for subscription registration
 
 **1.0.0** (2017-04-29)